OnTheSpot: introduce user hint prompt system
OnTheSpot uses a snackbar design to prompt the user for changes. This is
designed to be unobtrusive and fit in to the android system design.

Contributors:
@D4rKn3sSyS
@aaronpoweruser
@Anu6is
@carlosavignano
@signalv
@evan-a-a
@thecrazyskull

Signed-off-by: Carlo Savignano <carlosavignano@aospa.co>
Signed-off-by: SferaDev <alexisrico@aospa.co>
Signed-off-by: Evan Anderson <evananderson@aospa.co>

OnTheSpot: Rework the OTS framework

Drop the animations and the graphics and everything so that the dialog
just looks like it was part of the system. Ask Arz more about this
change if you don't believe in it. Additionally, make the "ask later"
option a simple "just once" to avoid confusion about whether the button
will or will not make the UI feature appear. Along the way make the
whole android.util.SettingConfirmationHelper ordeal more adaptable to
all kinds of possible situations and move code around here and there.

Rework of Change-Id: I725a8164c13c5381ddbc15fde29902a258c1ecc4

To be merged with Change-Id:
  packages/apps/Settings: Ia1742fc3d7dfa5af9e7f29f68dc8ec76eea5a98d

Change-Id: I4cc5b89c7fc96d8dcc1cd65e2a1e03aff8c8af39
Signed-off-by: Evan Anderson <evananderson@aospa.co>

OnTheSpot: Make the settings values public

Other classes will have to use the helper's constants if they do any
manual value checks. Re-defining the constants in each of those
classes is just prone to resulting in a messy codebase - it is better
to reference the original constants in all usage cases. This change
allows doing that. Nothing too complicated here.

Change-Id: I0f382107cc8166905d89a6b6605a282fc47dce07

Settings: Move SETTINGS_TO_RESET from System to Secure

Change-Id: I15e98feff4278998555ab49bc48aadc28f2ff6c9
Signed-off-by: Evan Anderson <evananderson@aospa.co>

OTS: Access settings in Secure and not System

Change-Id: I33a34a51f1a9ddbbbd586fd046b981817d218a9f

OnTheSpot: Switch to a snackbar

This commit moves OTS code over to SystemUI and makes it use a
snackbar which is less intrusive. <3

Change-Id: I772e577f63fbda25eb6050d56e581f3ebbb222a7
Signed-off-by: Evan Anderson <evananderson@aospa.co>

OnTheSpot: Add snackbar for devices with no navbar

The snackbar should not use the same checks the navbar does. This
commit removes that linking and sets the snackbar up even if the
navbar doesn't get set up.

Change-Id: Ib8d7e15ab17de8cb6433cdba3415148791875ead
Signed-off-by: Evan Anderson <evananderson@aospa.co>

OTS: Animate the snackbar only upon changes

Prior to this commit, each time the show() was called, the animation
took place. From this commit and onwards, the animation will only
happen when there are actual changes that will be visible to the
user (that is, the on-screen message has changed).

Change-Id: Iea577d93499f1c14928f506f77d56556b78e4b09

OTS: Clean up snackbar's logic

The main Handler instance that is used for scheduling the hiding calls
should be kept around so that the calls can be rescheduled later on.
Along that, generic CharSequences shouldn't be compared to equal one
another due to possible undefined behaviour. Instead, Strings should
be compared to each other, when deciding whether the animation should
take place.

Change-Id: Ia7f9847b6749c26aef9f2087724a0924c9218fe7
Signed-off-by: Evan Anderson <evananderson@aospa.co>
Signed-off-by: Carlo Savignano <carlosavignano@aospa.co>